refactor(web-ui): remove live terminal panel
