+++ tests/unit/web-ui-behavior-parity.test.mjs
@@ -336,15 +336,6 @@ test('captured bundled app skeleton only exposes expected data key drift versus 
         'webhookSaving',
         'webhookTestResult',
         'webhookTesting',
-        'terminalPanelOpen',
-        'terminalLines',
-        'terminalPaused',
-        'terminalSearchQuery',
-        'terminalCommandInput',
-        'terminalRunning',
-        'terminalSocket',
-        'terminalPendingBuffer',
-        'terminalMaxLines',
         'shareCommandPrefix',
         'sessionsUsageLoadedLimit',
         'taskOrchestrationTabEnabled',
@@ -378,15 +369,6 @@ test('captured bundled app skeleton only exposes expected data key drift versus 
         'webhookSaving',
         'webhookTestResult',
         'webhookTesting',
-        'terminalPanelOpen',
-        'terminalLines',
-        'terminalPaused',
-        'terminalSearchQuery',
-        'terminalCommandInput',
-        'terminalRunning',
-        'terminalSocket',
-        'terminalPendingBuffer',
-        'terminalMaxLines',
         'openCloneClaudeConfigModal',
         'openCloneProviderModal'
     ];
@@ -478,14 +460,6 @@ test('captured bundled app skeleton only exposes expected data key drift versus 
         'saveWebhookSettings',
         'testWebhook',
         'toggleWebhookEvent',
-        'toggleTerminalPanel',
-        'clearTerminalOutput',
-        'pauseTerminal',
-        'resumeTerminal',
-        'runTerminalCommand',
-        'killTerminalCommand',
-        '_appendTerminalLine',
-        '_flushTerminalBuffer',
         'restoreNavStateFromStorage',
         'cancelScheduledSessionListViewportFill',
         'canSubmitProvider',
