+++ web-ui/index.html
@@ -23,7 +23,6 @@
     <!-- @include ./partials/index/panel-settings.html -->
     <!-- @include ./partials/index/panel-market.html -->
     <!-- @include ./partials/index/panel-plugins.html -->
-    <!-- @include ./partials/index/panel-terminal.html -->
     <!-- @include ./partials/index/layout-footer.html -->
     <!-- @include ./partials/index/modals-basic.html -->
     <!-- @include ./partials/index/modal-openclaw-config.html -->
