+++ web-ui/modules/app.methods.index.mjs
@@ -29,7 +29,7 @@ import { createStartupClaudeMethods } from './app.methods.startup-claude.mjs';
 import { createSkillsMethods } from './skills.methods.mjs';
 import { createPluginsMethods } from './plugins.methods.mjs';
 import { createI18nMethods } from './i18n.mjs';
-import { createWebhookTerminalMethods } from './app.methods.webhook-terminal.mjs';
+import { createWebhookMethods } from './app.methods.webhook.mjs';
 import {
     CONFIG_MODE_SET,
     getProviderConfigModeMeta
@@ -44,7 +44,7 @@ import {
 export function createAppMethods() {
     return {
         ...createI18nMethods(),
-        ...createWebhookTerminalMethods(),
+        ...createWebhookMethods(),
         ...createStartupClaudeMethods({
             api,
             defaultModelContextWindow: DEFAULT_MODEL_CONTEXT_WINDOW,
