+++ web-ui/styles.css
@@ -19,4 +19,4 @@
 @import url('./styles/docs-panel.css');
 @import url('./styles/feedback.css');
 @import url('./styles/responsive.css');
-@import url('./styles/terminal-webhook.css');
+@import url('./styles/webhook.css');
